refactor: remove guava dependency
