+++ build.gradle.kts
@@ -31,7 +31,6 @@ subprojects {
     }
 
     dependencies {
-        api("com.google.guava:guava:30.1-jre")
         api("io.leangen.geantyref:geantyref:1.3.11")
 
         testImplementation("org.junit.jupiter:junit-jupiter:5.+")
