+++ core/src/main/java/net/kyori/moonshine/placeholder/ContinuanceValue.java
@@ -22,7 +22,7 @@ import io.leangen.geantyref.GenericTypeReflector;
 import io.leangen.geantyref.TypeToken;
 import java.lang.reflect.AnnotatedType;
 import java.lang.reflect.Type;
-import javax.annotation.concurrent.ThreadSafe;
+import net.kyori.moonshine.annotation.meta.ThreadSafe;
 import org.checkerframework.dataflow.qual.Pure;
 import org.checkerframework.dataflow.qual.SideEffectFree;
