+++ core/src/main/java/net/kyori/moonshine/util/Weighted.java
@@ -19,7 +19,7 @@
 package net.kyori.moonshine.util;
 
 import java.util.Objects;
-import javax.annotation.concurrent.ThreadSafe;
+import net.kyori.moonshine.annotation.meta.ThreadSafe;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.qual.Pure;
