+++ standard/src/main/java/net/kyori/moonshine/strategy/supertype/StandardSupertypeThenInterfaceSupertypeStrategy.java
@@ -27,7 +27,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.stream.Stream;
-import javax.annotation.concurrent.ThreadSafe;
+import net.kyori.moonshine.annotation.meta.ThreadSafe;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 /**
